requirements.txt
fastapi
uvicorn
ipykernel
pandas
numpy
tensorflow
keras
matplotlib
keras_preprocessing
pillow
python-multipart
Jinja2
mysql-connector-python
sqlalchemy
aiofiles
aiomysql
sqlalchemy
pymysql
cryptography